#55 Add Product drop down (Like user roles) Remove product pop up Add workflow and edit

+++ Msr.Services/PurchesOrder/Procedures/AddPurchaseOrderProcedure.cs
@@ -105,7 +105,7 @@ namespace Msr.Services.PurchesOrder.Procedures
         public string BillingEmail { get; set; }
 
         [StoredProcedureParameter(SqlDbType.Float, ParameterName = "TAX_RATE")]
-        public decimal? TaxRate { get; set; }
+        public double? TaxRate { get; set; }
 
         [StoredProcedureParameter(SqlDbType.VarChar, Size = 50, ParameterName = "StrNTLogin")]
         public string Strntlogin { get; set; }
